backend/requirements.txt
fastapi
httpx
pydantic-settings
uvicorn[standard]
yt-dlp
bgutil-ytdlp-pot-provider
pypresence
pillow
